FIX-2475 Command palette with better ux flow for query editor
