+++ Src/WitsmlExplorer.Frontend/components/ContentViews/QueryView/components/QueryOptions/TemplatePicker/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TemplatePicker.tsx";
