+++ Src/WitsmlExplorer.Frontend/components/ContentViews/QueryView/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./QueryView.tsx";
